Added zipfian distribution

+++ src/include/configuration.h
@@ -30,7 +30,7 @@ class configuration {
   size_t batch_size;
 
   // join selectivity threshold
-  int join_selectivity_threshold;
+  double join_selectivity_threshold;
 
   // Verbose output
   bool verbose;
